Refactored Merge Sort: fixed filename, added dynamic memory, and pure text docs
